Make minifiers work
While it is somewhat attractive to try and pass the minifier to the
minified Resource object as its compiler, that would necessitate for
the compiler to be able to change the resource's filename, which is
kludgy.

Thus, the clear separation between Compiler and Minifier we have now
is deemed more attractive, especially since a) the current code in
Resource is not too kludgy and b) we would need the mode_parent anyway
(since the parent might have to be compiled before minifying).

--HG--
branch : wosc-compilers